Revert "refactor: renamed all package root files to index"
This reverts commit a52e0b1599b7b9e98c069628413dc89d880206a9.
