+++ packages/everything/package.json
@@ -1,6 +1,7 @@
 {
   "name": "@functional-things/everything",
   "version": "1.0.0-rc.1",
+  "private": false,
   "description": "TypeScript utilities for functional programming.",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -46,10 +47,5 @@
   "jest": {
     "preset": "ts-jest",
     "testEnvironment": "node"
-  },
-  "dependencies": {
-    "@functional-things/compose": "workspace:packages/compose",
-    "@functional-things/maybe": "workspace:packages/maybe",
-    "@functional-things/optics": "workspace:packages/optics"
   }
 }
