+++ packages/lists/package.json
@@ -2,8 +2,8 @@
   "name": "@functional-things/lists",
   "version": "1.0.0-rc.1",
   "description": "The lists package from @functional-things.",
-  "main": "dist/index",
-  "types": "dist/index",
+  "main": "dist/lists.js",
+  "types": "dist/lists.d.ts",
   "files": [
     "dist/*"
   ],
