+++ packages/maybe/src/maybe.ts
